Leave Credit Table Update

+++ database/migrations/2025_02_27_135317_create_leave_credits_table.php
@@ -1,3 +1,5 @@
+
+
 <?php
 
 use Illuminate\Database\Migrations\Migration;
@@ -18,8 +20,8 @@ class CreateLeaveCreditsTable extends Migration
             $table->unsignedBigInteger('client_id');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('application_type_id');
-            $table->integer('number');
-            $table->integer('used');
+            $table->decimal('number', 8, 2);
+            $table->decimal('used', 8, 2);
             $table->timestamps();
             $table->softDeletes();
